Refactor ec.py module to lib/
