+++ demo2.py
@@ -8,7 +8,7 @@ import datetime
 import os
 import random
 
-from ec import commandlineArguments, ecIterator
+from lib.ec import commandlineArguments, ecIterator
 from grammar import Grammar
 from program import Primitive
 from lib.task import Task
