+++ logo.py
@@ -1,4 +1,4 @@
-from ec import ecIterator, commandlineArguments
+from lib.ec import ecIterator, commandlineArguments
 from grammar import Grammar
 from utilities import eprint, testTrainSplit, numberOfCPUs, loadPickle
 from lib.domains.logo.makeLogoTasks import makeTasks, montageTasks, drawLogo
