+++ text.py
@@ -1,4 +1,4 @@
-from ec import commandlineArguments, ecIterator
+from lib.ec import commandlineArguments, ecIterator
 from lib.domains.text.makeTextTasks import makeTasks, loadPBETasks
 from lib.domains.text.textPrimitives import primitives
 from lib.domains.list.listPrimitives import bootstrapTarget
