+++ tower.py
@@ -1,6 +1,6 @@
-from ec import *
+from lib.ec import *
 
-from lib.domains.tower.towerPrimitives import primitives, new_primitives, executeTower, animateTower
+from lib.domains.tower.towerPrimitives import primitives, new_primitives, animateTower
 from lib.domains.tower.makeTowerTasks import *
 from tower_common import renderPlan, towerLength, centerTower
 from utilities import *
